refactor(client): hooks check for session, no hocs
