+++ apps/notes/client/pages/login.tsx
@@ -1,6 +1,5 @@
-import { withoutAuth } from '@composer';
 import dynamic from 'next/dynamic';
 
 const LoginPage = dynamic(() => import('@pages').then((mod) => mod.LoginPage));
 
-export default withoutAuth(LoginPage);
+export default LoginPage;
