+++ apps/notes/client/pages/note.tsx
@@ -1,6 +1,5 @@
-import { withAuth } from '@composer';
 import dynamic from 'next/dynamic';
 
 const Note = dynamic(() => import('@pages').then((mod) => mod.Note));
 
-export default withAuth(Note);
+export default Note;
