+++ apps/notes/client/pages/register.tsx
@@ -1,8 +1,7 @@
-import { withoutAuth } from '@composer';
 import dynamic from 'next/dynamic';
 
 const RegisterPage = dynamic(() =>
     import('@pages').then((mod) => mod.RegisterPage)
 );
 
-export default withoutAuth(RegisterPage);
+export default RegisterPage;
