--- libs/composer/.babelrc
@@ -1,4 +1,0 @@
-{
-    "presets": ["@nx/next/babel"],
-    "plugins": []
-}
