--- libs/composer/README.md
@@ -1,7 +1,0 @@
-# composer
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test composer` to execute the unit tests via [Jest](https://jestjs.io).
