--- libs/composer/jest.config.ts
@@ -1,13 +1,0 @@
-/* eslint-disable */
-export default {
-    displayName: 'composer',
-    preset: '../../jest.preset.js',
-    transform: {
-        '^.+\\.[tj]sx?$': [
-            '@swc/jest',
-            { jsc: { transform: { react: { runtime: 'automatic' } } } },
-        ],
-    },
-    moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-    coverageDirectory: '../../coverage/libs/composer',
-};
