--- libs/composer/src/index.ts
@@ -1,2 +1,0 @@
-export * from './lib/with-auth/with-auth';
-export * from './lib/without-auth/without-auth';
