--- libs/composer/tsconfig.lib.json
@@ -1,23 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "compilerOptions": {
-        "outDir": "../../dist/out-tsc",
-        "types": ["node"]
-    },
-    "files": [
-        "../../node_modules/@nx/react/typings/cssmodule.d.ts",
-        "../../node_modules/@nx/next/typings/image.d.ts"
-    ],
-    "exclude": [
-        "jest.config.ts",
-        "src/**/*.spec.ts",
-        "src/**/*.test.ts",
-        "src/**/*.spec.tsx",
-        "src/**/*.test.tsx",
-        "src/**/*.spec.js",
-        "src/**/*.test.js",
-        "src/**/*.spec.jsx",
-        "src/**/*.test.jsx"
-    ],
-    "include": ["src/**/*.js", "src/**/*.jsx", "src/**/*.ts", "src/**/*.tsx"]
-}
