+++ tsconfig.base.json
@@ -18,7 +18,6 @@
             "@api": ["libs/api/src/index.ts"],
             "@auth": ["libs/auth/src/index.ts"],
             "@components": ["libs/components/src/index.ts"],
-            "@composer": ["libs/composer/src/index.ts"],
             "@config": ["libs/config/src/index.ts"],
             "@pages": ["libs/pages/src/index.ts"],
             "@prisma": ["libs/ts-prisma/src/index.ts"],
